refactor(structure): move page components to src/common/pages
Reorganizes project structure by creating a dedicated pages folder for
page components served by routes, separating them from reusable UI components.

Changes:
- Create src/common/pages/ directory
- Move all page components from components/ to pages/:
  * HomePage, PostsPage, PostDetailPage
  * UsersPage, UserProfilePage
  * NotFound, ServerError
- Update router imports in src/client/router.jsx
- Update router imports in src/server/router.jsx
- Fix relative imports in moved pages for Page and ui components
- Update README.md project structure section
- Update CLAUDE.md with new pages folder documentation

Benefits:
- Clear separation between pages (route targets) and components (reusables)
- Improved project organization and maintainability
- Easier to locate and work with page-level code
