+++ src/common/pages/PostDetailPage/PostDetailPage.stories.jsx
