+++ src/common/pages/PostsPage/PostsPage.stories.jsx
