+++ src/common/pages/PostsPage/index.js
